requirements.txt
apiclient==1.0.4
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
google-api-core==2.8.1
google-api-python-client==2.51.0
google-auth==2.7.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.2
gspread==5.4.0
httplib2==0.20.4
oauth2client==4.1.3
oauthlib==3.2.0
psycopg2==2.9.3
PyGreSQL==5.2.4
pyparsing==3.0.9
python-dotenv==0.20.0
requests==2.28.0
requests-oauthlib==1.3.1
urllib3==1.26.9
